content/xul/document/test/test_bug335375.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<?xul-overlay href="overlay1_bug335375.xul"?>
<?xul-overlay href="overlay2_bug335375.xul"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=335375
-->
<window title="Mozilla Bug 335375"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="RunTest();">
  <title>Mozilla Bug 335375</title>

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>      

  <script type="application/javascript">
    <![CDATA[
      SimpleTest.waitForExplicitFinish();

      function RunTest()
      {
        var cmd = document.getElementById("cmd_test");
        var button = document.getElementById("button_test");

        is(cmd.getAttribute("accesskey"), "C",
           "checking command has an accesskey");
        is(button.getAttribute("accesskey"), cmd.getAttribute("accesskey"),
           "checking command and button have the same accesskey");

        cmd.setAttribute("accesskey", "D");
        is(button.getAttribute("accesskey"), "D",
           "checking button has inherited new accesskey from command");

        SimpleTest.finish();
      }
    ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=335375"
       target="_blank">Mozilla Bug 335375</a>
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>
  </body>

  <commandset id="test_commandset"/>
  <toolbarbutton id="button_test"/>

</window>
